+++ examples/rest_server.cc
@@ -39,7 +39,7 @@ public:
 
     void init(size_t thr = 2) {
         auto opts = Http::Endpoint::options()
-            .threads(thr);
+            .threads(static_cast<int>(thr));
         httpEndpoint->init(opts);
         setupRoutes();
     }
@@ -150,10 +150,10 @@ int main(int argc, char *argv[]) {
     int thr = 2;
 
     if (argc >= 2) {
-        port = std::stol(argv[1]);
+        port = static_cast<uint16_t>(std::stol(argv[1]));
 
         if (argc == 3)
-            thr = std::stol(argv[2]);
+            thr = std::stoi(argv[2]);
     }
 
     Address addr(Ipv4::any(), port);
