+++ tests/rest_server_test.cc
@@ -22,7 +22,7 @@ public:
       : httpEndpoint(std::make_shared<Http::Endpoint>(addr)) {}
 
   void init(size_t thr = 2) {
-    auto opts = Http::Endpoint::options().threads(thr);
+    auto opts = Http::Endpoint::options().threads(static_cast<int>(thr));
     httpEndpoint->init(opts);
     setupRoutes();
   }
